+++ src/Elements/Detail.js
@@ -319,7 +319,7 @@ export default class Detail {
       .on('click', c('.back'), this.hide)
       .on('click', c('.refresh'), () => {
         this._render()
-        devtools.notify('Refreshed')
+        devtools.notify('Refreshed', { icon: 'success' })
       })
   }
   _initObserver() {