+++ test/elements.js
@@ -1,14 +1,13 @@
 describe('elements', function () {
   let tool = eruda.get('elements')
-  let $tool = $('.eruda-elements')
 
   beforeEach(function () {
     eruda.show('elements')
   })
 
   describe('api', function () {
-    it('set element', function () {
-      tool.set(document.body)
+    it('select element', function () {
+      tool.select(document.body)
     })
   })
 })